+++ lib/noticed/delivery_methods/base.rb
@@ -4,12 +4,13 @@ module Noticed
       extend ActiveModel::Callbacks
       define_model_callbacks :deliver
 
-      attr_reader :notification, :options, :recipient
+      attr_reader :notification, :options, :recipient, :record
 
       def perform(notification_class:, options:, params:, recipient:, record:)
         @notification = notification_class.constantize.new(params)
         @options = options
         @recipient = recipient
+        @record = record
 
         # Make notification aware of database record and recipient during delivery
         @notification.record = record